Refactor schema and resolvers to be testable
